refactor: 拆分 GenUI schema 职责
