+++ src/client/genui-runtime/schema.ts
@@ -53,14 +53,6 @@ export interface ComponentSchema {
   readonly validator?: ComponentValidatorMetadata
 }
 
-export interface GenuiDiagnostic {
-  readonly kind: 'alias' | 'unknown-field'
-  readonly path: string
-  readonly message: string
-  readonly type?: string
-  readonly field?: string
-  readonly canonical?: string
-}
 
 /** Canonical enum domains shared by schema validation and repair. */
 export const TEXT_SIZES = ['h1', 'h2', 'h3', 'body', 'muted', 'caption'] as const
@@ -297,234 +289,3 @@ export const COMPONENT_SCHEMAS: Readonly<Record<string, ComponentSchema>> = {
 } as const
 
 export const GENUI_NATIVE_TYPES: ReadonlySet<string> = new Set(Object.keys(COMPONENT_SCHEMAS))
-
-function record(value: unknown): Record<string, unknown> | undefined {
-  return typeof value === 'object' && value !== null && !Array.isArray(value)
-    ? value as Record<string, unknown>
-    : undefined
-}
-
-function isNode(value: unknown): value is Record<string, unknown> {
-  const candidate = record(value)
-  return candidate !== undefined && typeof candidate.type === 'string'
-}
-
-function normalizeAliasFields(value: Record<string, unknown>, path: string, type: string, warnings: GenuiDiagnostic[]): Record<string, unknown> {
-  const definition = COMPONENT_SCHEMAS[type]
-  if (definition === undefined) return value
-  const out: Record<string, unknown> = { ...value }
-  for (const [alias, canonical] of Object.entries(definition.aliases)) {
-    if (!(alias in out)) continue
-    const aliasPath = `${path}.${alias}`
-    const keptCanonical = canonical in out
-    if (!keptCanonical) out[canonical] = out[alias]
-    delete out[alias]
-    warnings.push({
-      kind: 'alias',
-      path: aliasPath,
-      message: keptCanonical
-        ? `${aliasPath} is ignored because canonical field '${canonical}' is present`
-        : `${aliasPath} normalized/adopted as '${canonical}'`,
-      type,
-      field: alias,
-      canonical,
-    })
-  }
-  return out
-}
-
-function normalizeNode(value: unknown, path: string, warnings: GenuiDiagnostic[]): unknown {
-  if (!isNode(value)) return value
-  const type = value.type as string
-  const definition = COMPONENT_SCHEMAS[type]
-  // Custom nodes are opaque by contract: don't inspect or rewrite their data.
-  if (definition === undefined) return value
-  const out = normalizeAliasFields(value, path, type, warnings)
-  const normalizeNodeValue = (child: unknown, childPath: string): unknown => normalizeNode(child, childPath, warnings)
-  const normalizeNodeArray = (children: unknown, childPath: string): unknown => Array.isArray(children)
-    ? children.map((child, index) => normalizeNodeValue(child, `${childPath}[${index}]`))
-    : children
-
-  if (type === 'row' || type === 'col' || type === 'grid' || type === 'card' || type === 'file-tree' || type === 'timeline' || type === 'breadcrumb') {
-    if (type !== 'file-tree' && type !== 'timeline' && type !== 'breadcrumb') out.items = normalizeNodeArray(out.items, `${path}.items`)
-  } else if (type === 'list' && Array.isArray(out.items)) {
-    out.items = out.items.map((child, index) => isNode(child) ? normalizeNodeValue(child, `${path}.items[${index}]`) : child)
-  } else if (type === 'tabs' && Array.isArray(out.tabs)) {
-    out.tabs = out.tabs.map((tab, index) => {
-      const holder = record(tab)
-      if (holder === undefined) return tab
-      const normalizedHolder = { ...holder }
-      if ('content' in normalizedHolder) {
-        const tabPath = `${path}.tabs[${index}].content`
-        const hasItems = 'items' in normalizedHolder
-        if (!hasItems) normalizedHolder.items = normalizedHolder.content
-        delete normalizedHolder.content
-        warnings.push({
-          kind: 'alias',
-          path: tabPath,
-          message: hasItems
-            ? `${tabPath} is ignored because canonical field 'items' is present`
-            : `${tabPath} normalized/adopted as 'items'`,
-          type,
-          field: 'content',
-          canonical: 'items',
-        })
-      }
-      normalizedHolder.items = Array.isArray(normalizedHolder.items)
-        ? normalizedHolder.items.map((child, childIndex) => normalizeNodeValue(child, `${path}.tabs[${index}].items[${childIndex}]`))
-        : normalizedHolder.items === undefined
-          ? normalizedHolder.items
-          : [normalizeNodeValue(normalizedHolder.items, `${path}.tabs[${index}].items[0]`)]
-      return normalizedHolder
-    })
-  } else if (type === 'accordion' && Array.isArray(out.items)) {
-    out.items = out.items.map((item, index) => {
-      const holder = record(item)
-      if (holder === undefined) return item
-      return { ...holder, items: Array.isArray(holder.items) ? holder.items.map((child, childIndex) => normalizeNodeValue(child, `${path}.items[${index}].items[${childIndex}]`)) : holder.items }
-    })
-  }
-  return out
-}
-
-/**
- * Normalize a raw GenUI value into canonical field names.
- *
- * Only deterministic aliases and structural aliases are changed. Resource
- * limits, type repair, security filtering, and semantic validation remain in
- * the guard layer. Unknown component types are returned opaque.
- *
- * @param value - Raw GenUI spec or bare native component.
- * @returns Canonical value and stable alias diagnostics.
- */
-export function normalizeGenuiSpec(value: unknown): { value: unknown; warnings: GenuiDiagnostic[] } {
-  const warnings: GenuiDiagnostic[] = []
-  const root = record(value)
-  if (root === undefined) return { value, warnings }
-  const out = { ...root }
-  if (Array.isArray(out.items)) {
-    out.items = out.items.map((item, index) => normalizeNode(item, `items[${index}]`, warnings))
-  } else if (typeof out.type === 'string') {
-    return { value: normalizeNode(out, 'spec', warnings), warnings }
-  }
-  return { value: out, warnings }
-}
-
-function visitNativeNodes(value: unknown, path: string, visit: (node: Record<string, unknown>, path: string, schema: ComponentSchema) => void): void {
-  if (!isNode(value)) return
-  const type = value.type as string
-  const definition = COMPONENT_SCHEMAS[type]
-  if (definition === undefined) return
-  visit(value, path, definition)
-  const children = (child: unknown, childPath: string): void => visitNativeNodes(child, childPath, visit)
-  if ((type === 'row' || type === 'col' || type === 'grid' || type === 'card') && Array.isArray(value.items)) {
-    value.items.forEach((child, index) => children(child, `${path}.items[${index}]`))
-  } else if (type === 'list' && Array.isArray(value.items)) {
-    value.items.forEach((child, index) => children(child, `${path}.items[${index}]`))
-  } else if (type === 'tabs' && Array.isArray(value.tabs)) {
-    value.tabs.forEach((tab, index) => {
-      const holder = record(tab)
-      if (holder === undefined) return
-      if (Array.isArray(holder.items)) holder.items.forEach((child, childIndex) => children(child, `${path}.tabs[${index}].items[${childIndex}]`))
-      else children(holder.items, `${path}.tabs[${index}].items`)
-    })
-  } else if (type === 'accordion' && Array.isArray(value.items)) {
-    value.items.forEach((item, index) => {
-      const holder = record(item)
-      if (holder?.items !== undefined && Array.isArray(holder.items)) holder.items.forEach((child, childIndex) => children(child, `${path}.items[${index}].items[${childIndex}]`))
-    })
-  }
-}
-
-function pushUnknownField(
-  warnings: GenuiDiagnostic[],
-  path: string,
-  field: string,
-  type: string,
-): void {
-  warnings.push({
-    kind: 'unknown-field',
-    path: `${path}.${field}`,
-    message: `${path}.${field}: unknown field for '${type}'`,
-    type,
-    field,
-  })
-}
-
-function diagnoseRecordFields(
-  value: unknown,
-  path: string,
-  definition: ComponentRecordSchema,
-  type: string,
-  warnings: GenuiDiagnostic[],
-): void {
-  const holder = record(value)
-  if (holder === undefined) return
-  for (const field of Object.keys(holder)) {
-    if (field in definition.fields) continue
-    pushUnknownField(warnings, path, field, type)
-  }
-  for (const [field, nested] of Object.entries(definition.nested)) {
-    const nestedValue = holder[field]
-    if (Array.isArray(nestedValue)) {
-      nestedValue.forEach((item, index) => diagnoseRecordFields(item, `${path}.${field}[${index}]`, nested, type, warnings))
-    } else if (nestedValue !== undefined) {
-      diagnoseRecordFields(nestedValue, `${path}.${field}`, nested, type, warnings)
-    }
-  }
-}
-
-function diagnoseNestedFields(
-  node: Record<string, unknown>,
-  path: string,
-  definition: ComponentSchema,
-  warnings: GenuiDiagnostic[],
-): void {
-  for (const [field, nested] of Object.entries(definition.nested)) {
-    const nestedValue = node[field]
-    if (Array.isArray(nestedValue)) {
-      nestedValue.forEach((item, index) => diagnoseRecordFields(item, `${path}.${field}[${index}]`, nested, node.type as string, warnings))
-    } else if (nestedValue !== undefined) {
-      diagnoseRecordFields(nestedValue, `${path}.${field}`, nested, node.type as string, warnings)
-    }
-  }
-}
-
-/**
- * Diagnose unknown direct fields on native nodes.
- *
- * Unknown types are intentionally skipped so custom renderers retain their
- * opaque extension payloads. Native unknown fields are warnings, not errors.
- *
- * @param value - Canonical or raw GenUI value.
- * @returns Stable field diagnostics in tree order.
- */
-export function diagnoseUnknownGenuiFields(value: unknown): GenuiDiagnostic[] {
-  const warnings: GenuiDiagnostic[] = []
-  const root = record(value)
-  if (root === undefined) return warnings
-  const visit = (node: Record<string, unknown>, path: string, definition: ComponentSchema): void => {
-    for (const field of Object.keys(node)) {
-      if (field === 'type' || field in definition.fields) continue
-      pushUnknownField(warnings, path, field, node.type as string)
-    }
-    diagnoseNestedFields(node, path, definition, warnings)
-  }
-  if (Array.isArray(root.items)) {
-    for (const field of Object.keys(root)) {
-      if (field in GENUI_SPEC_SCHEMA.fields) continue
-      pushUnknownField(warnings, 'spec', field, 'spec')
-    }
-    root.items.forEach((item, index) => visitNativeNodes(item, `items[${index}]`, visit))
-  } else if (typeof root.type === 'string') {
-    // A bare component root is a documented shorthand, so its `type` belongs
-    // to the native node schema rather than the root specification schema.
-    visitNativeNodes(root, 'spec', visit)
-  } else {
-    for (const field of Object.keys(root)) {
-      if (field in GENUI_SPEC_SCHEMA.fields) continue
-      pushUnknownField(warnings, 'spec', field, 'spec')
-    }
-  }
-  return warnings
-}
